+++ includes/classes/Locations.php
@@ -28,17 +28,17 @@ class ad_skip_hire_locations
     public function location_post_type() 
     {
         $labels = [
-            'name'                  => _x( 'Delivery Locations', $this->cpt_prefix),
-            'singular_name'         => _x( 'Delivery Location', $this->cpt_prefix),
+            'name'                  => _x( 'Locations', $this->cpt_prefix),
+            'singular_name'         => _x( 'Location', $this->cpt_prefix),
             'add_new'               => _x( 'Add New', $this->cpt_prefix),
-            'add_new_item'          => _x( 'Add New Delivery Location', $this->cpt_prefix),
-            'edit_item'             => _x( 'Edit Delivery Location', $this->cpt_prefix ),
-            'new_item'              => _x( 'New Delivery Location', $this->cpt_prefix),
-            'view_item'             => _x( 'View Delivery Location', $this->cpt_prefix),
-            'search_items'          => _x( 'Search Delivery Location', $this->cpt_prefix ),
-            'not_found'             => _x( 'No Delivery Locations found', $this->cpt_prefix ),
-            'not_found_in_trash'    => _x( 'No Delivery Locations found in Trash', $this->cpt_prefix ),
-            'menu_name'             => _x( 'Delivery Locations', $this->cpt_prefix ),
+            'add_new_item'          => _x( 'Add New Location', $this->cpt_prefix),
+            'edit_item'             => _x( 'Edit Location', $this->cpt_prefix ),
+            'new_item'              => _x( 'New Location', $this->cpt_prefix),
+            'view_item'             => _x( 'View Location', $this->cpt_prefix),
+            'search_items'          => _x( 'Search Location', $this->cpt_prefix ),
+            'not_found'             => _x( 'No Locations found', $this->cpt_prefix ),
+            'not_found_in_trash'    => _x( 'No Locations found in Trash', $this->cpt_prefix ),
+            'menu_name'             => _x( 'Locations', $this->cpt_prefix ),
         ];
      
         $args = [