+++ includes/views/orderConfirmation.php
@@ -1,4 +1,4 @@
-<?php if($_GET['success'] == true): ?>
+<?php if($_GET['success'] == 'true'): ?>
     <p>Thank you for your order.</p>
 <?php else: ?>
     <p>Sorry, there was a problem with processing your payment.</p>