+++ skip-hire.php
@@ -489,7 +489,7 @@ class ad_skip_hire
             $subTotal = $permit['price'] + $skip['price'];
 
             # Coupon
-            if( isset($_POST['ash_coupon'] ) ) {
+            if( isset( $_POST['ash_coupon'] ) && $_POST['ash_coupon'] != null ) {
 
                 $coupons = new WP_Query([
                     'post_type'         => 'ash_coupons',
@@ -594,6 +594,7 @@ class ad_skip_hire
     {
         $this->paypal->authorised_payment_check();
 
+        add_post_meta( $postID, 'ash_orders_status', 'complete');
         add_post_meta( $_SESSION['ash_order_id'], 'ash_orders_paypal_id', $_GET['paymentId']);
         add_post_meta( $_SESSION['ash_order_id'], 'ash_orders_paypal_payer_id', $_GET['PayerID']);
     }