+++ CHANGELOG.md
@@ -3,8 +3,12 @@
 ## Version 0.3
 
 - [Bug] Integrated a regex check for the post code submit to force a post code and stop the form breaking later on.
+- [Bug] Fixing coupon being added 100% of time.
+- [Bug] Fixing PayPal payment update.
 - [Improvement] Adding a settings panel to allow the user to manage settings.
 - [Improvement] Cleaning up menu names to make them more readable.
+- [Improvement] Improved order overview for quoting when on the phone.
+- [Improvement] Tweaks to wording to make it clearer for the user flow.
 
 ## Version 0.2
 - [Improvement] Restructuring filebase for easier use, declarations in classes as required.