+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
-## Version 0.3
+## Version 0.4
 
+- [Bug] Fixing catch for if no waste was selected on the form.
+
+## Version 0.3
 - [Bug] Integrated a regex check for the post code submit to force a post code and stop the form breaking later on.
 - [Bug] Fixing coupon being added 100% of time.
 - [Bug] Fixing PayPal payment update.